Fix Dependencies and Tests for filter
